rename aggregateFunction to AggregateFunction
