+++ src/AggregateFunction.ts
